// web/src/app/api/generated/src/models/CloseoutContext.ts
export interface CloseoutContext {
  /**
   *
   * @type {string}
   * @memberof CloseoutContext
   */
  appointmentId?: string;
  /**
   *
   * @type {boolean}
   * @memberof CloseoutContext
   */
  canComplete?: boolean;
  /**
   *
   * @type {Date}
   * @memberof CloseoutContext
   */
  endsAt?: Date;
  /**
   *
   * @type {ExecutionSummary}
   * @memberof CloseoutContext
   */
  execution?: ExecutionSummary;
  /**
   *
   * @type {boolean}
   * @memberof CloseoutContext
   */
  executionRecorded?: boolean;
  /**
   *
   * @type {boolean}
   * @memberof CloseoutContext
   */
  executionRequired?: boolean;
  /**
   *
   * @type {Array<MeasurementFact>}
   * @memberof CloseoutContext
   */
  measurements?: Array<MeasurementFact>;
  /**
   *
   * @type {string}
   * @memberof CloseoutContext
   */
  participantId?: string;
  /**
   *
   * @type {string}
   * @memberof CloseoutContext
   */
  participantName?: string;
  /**
   *
   * @type {string}
   * @memberof CloseoutContext
   */
  shortPurpose?: string;
  /**
   *
   * @type {Date}
   * @memberof CloseoutContext
   */
  startsAt?: Date;
  /**
   *
   * @type {string}
   * @memberof CloseoutContext
   */
  status?: string;
  /**
   *
   * @type {SummaryNote}
   * @memberof CloseoutContext
   */
  summaryNote?: SummaryNote;
  /**
   *
   * @type {number}
   * @memberof CloseoutContext
   */
  version?: number;
}