// web/src/app/api/generated/src/models/ParticipantMeasurementCommand.ts
export interface ParticipantMeasurementCommand {
  /**
   *
   * @type {string}
   * @memberof ParticipantMeasurementCommand
   */
  activity?: string;
  /**
   *
   * @type {string}
   * @memberof ParticipantMeasurementCommand
   */
  appointmentId?: string;
  /**
   *
   * @type {string}
   * @memberof ParticipantMeasurementCommand
   */
  bodyArea?: string;
  /**
   *
   * @type {string}
   * @memberof ParticipantMeasurementCommand
   */
  customLabel?: string;
  /**
   *
   * @type {string}
   * @memberof ParticipantMeasurementCommand
   */
  exercise?: string;
  /**
   *
   * @type {Date}
   * @memberof ParticipantMeasurementCommand
   */
  measuredAt?: Date;
  /**
   *
   * @type {string}
   * @memberof ParticipantMeasurementCommand
   */
  note?: string;
  /**
   *
   * @type {ParticipantMeasurementCommandPresetIdEnum}
   * @memberof ParticipantMeasurementCommand
   */
  presetId?: ParticipantMeasurementCommandPresetIdEnum;
  /**
   *
   * @type {string}
   * @memberof ParticipantMeasurementCommand
   */
  unit?: string;
  /**
   *
   * @type {number}
   * @memberof ParticipantMeasurementCommand
   */
  value?: number;
}

/**
 * @export
 */
export const ParticipantMeasurementCommandPresetIdEnum = {
  BodyWeight: 'BODY_WEIGHT',
  BodyCircumference: 'BODY_CIRCUMFERENCE',
  MaxLoad: 'MAX_LOAD',
  Distance: 'DISTANCE',
  CompletionTime: 'COMPLETION_TIME',
  HoldDuration: 'HOLD_DURATION',
  Custom: 'CUSTOM',
} as const;
export type ParticipantMeasurementCommandPresetIdEnum =
  (typeof ParticipantMeasurementCommandPresetIdEnum)[keyof typeof ParticipantMeasurementCommandPresetIdEnum];